+++ src/pipeline.py
@@ -1,10 +1,9 @@
 """The main pipeline"""
 # Core Python modules
-import time
 import logging
 
 # Our local modules
-from src.utils import runlog_refactor as runlog
+from src.utils import runlog
 from src._version import __version__ as version
 from src.utils.config import validate_config, merge_configs
 from src.utils.wrappers import logger_creator
